Refactor the result handler (#171)
* 1. Refactor the result handler
* 2. upgrade dependencies
* 3. Fix unexpected behavior of connection without spring environment
* 4. CI： update workflow config file
* 5. update pom

+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/support/R2dbcStatementLog.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2009-2024 the original author or authors.
+ *    Copyright 2009-2025 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -80,7 +80,7 @@ public class R2dbcStatementLog {
      *
      * @param rows the rows
      */
-    public void logTotal(Integer rows) {
+    public void logTotal(Long rows) {
         debug("     Total: " + rows, false);
     }
